jetson_server/app2/requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.35.0
pydantic==2.11.3
python-multipart==0.0.20
pillow==10.4.0
opencv-python
ultralytics==8.3.82